# openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the notification server API that handles both external user notification subscriptions and internal notifications",
        "title": "Notification Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "GNU General Public License v3.0",
            "url": "https://github.com/lz1marine/notification-service/?tab=GPL-3.0-1-ov-file"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/internal/notifications/{id}": {
            "post": {
                "description": "post a notification to a channel given an event id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Post a notification to a channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The request body.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ChannelNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ChannelNotificationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications": {
            "get": {
                "description": "gets the list of all channels, including whether they are enabled",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Gets the list of all channels, including whether they are enabled",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ChannelResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/sub/{id}": {
            "get": {
                "description": "gets the list of all channels subscribed to by the user, including whether they are enabled for the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Gets the list of all channels subscribed to by the user, including whether they are enabled for the user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ChannelResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "patch he channel list that the user has subscribed to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Patch the channel list that the user has subscribed to",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The request body.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SetChannelsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.SetChannelsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "v1.Channel": {
            "type": "object",
            "properties": {
                "is_enabled": {
                    "description": "IsEnabled is a flag indicating if the channel is active",
                    "type": "boolean"
                },
                "name": {
                    "description": "Name is the name of the channel",
                    "type": "string"
                }
            }
        },
        "v1.ChannelNotificationRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "description": "Channel is the name of the channel",
                    "type": "string"
                },
                "message": {
                    "description": "Message is the content of the message",
                    "type": "string"
                },
                "subject": {
                    "description": "Subject is the subject of the message",
                    "type": "string"
                },
                "template_id": {
                    "description": "TemplateID is the id of the template",
                    "type": "string"
                },
                "topic_id": {
                    "description": "Topic is the id of the topic",
                    "type": "string"
                }
            }
        },
        "v1.ChannelNotificationResponse": {
            "type": "object"
        },
        "v1.ChannelResponse": {
            "type": "object",
            "properties": {
                "channels": {
                    "description": "Channels is an array of channel objects",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Channel"
                    }
                }
            }
        },
        "v1.SetChannelsRequest": {
            "type": "object",
            "properties": {
                "channels": {
                    "description": "Channels is an array of channel objects",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Channel"
                    }
                }
            }
        },
        "v1.SetChannelsResponse": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "basic"
        }
    }
}